# backend/docs/swagger.json
{
  "basePath": "/api/v1",
  "definitions": {
...
    "models.APIResponse": {
      "properties": {
        "data": {},
        "message": {
          "example": "Operation successful",
          "type": "string"
        },
        "status": {
          "example": "success",
          "type": "string"
        }
      },
      "type": "object"
    },
...
    "services.PaymentCreateRequest": {
      "properties": {
        "allocations": {
          "items": {
            "$ref": "#/definitions/services.InvoiceAllocation"
          },
          "type": "array"
        },
        "amount": {
          "minimum": 0,
          "type": "number"
        },
        "bill_allocations": {
          "items": {
            "$ref": "#/definitions/services.BillAllocation"
          },
          "type": "array"
        },
        "cash_bank_id": {
          "type": "integer"
        },
        "contact_id": {
          "type": "integer"
        },
        "date": {
          "type": "string"
        },
        "method": {
          "type": "string"
        },
        "notes": {
          "type": "string"
        },
        "reference": {
          "type": "string"
        }
      },
      "required": [
        "amount",
        "contact_id",
        "date",
        "method"
      ],
      "type": "object"
    },
...
  },
  "host": "localhost:8080",
  "info": {
    "contact": {
      "email": "support@swagger.io",
      "name": "API Support",
      "url": "http://www.swagger.io/support"
    },
    "description": "API untuk aplikasi sistem akuntansi yang komprehensif dengan fitur lengkap manajemen keuangan, inventory, sales, purchases, dan reporting.",
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "termsOfService": "http://swagger.io/terms/",
    "title": "Sistema Akuntansi API",
    "version": "1.0"
  },
  "paths": {
...
    "/api/payments": {
      "get": {
        "consumes": [
          "application/json"
        ],
        "deprecated": true,
        "description": "DEPRECATED: This endpoint may cause double posting. Use SSOT Payment routes instead. Get paginated list of payments with filters",
        "parameters": [
          {
            "description": "Page number",
            "in": "query",
            "name": "page",
            "type": "integer"
          },
          {
            "description": "Items per page",
            "in": "query",
            "name": "limit",
            "type": "integer"
          },
          {
            "description": "Filter by contact ID",
            "in": "query",
            "name": "contact_id",
            "type": "integer"
          },
          {
            "description": "Filter by status",
            "in": "query",
            "name": "status",
            "type": "string"
          },
          {
            "description": "Filter by payment method",
            "in": "query",
            "name": "method",
            "type": "string"
          },
          {
            "description": "Start date (YYYY-MM-DD)",
            "in": "query",
            "name": "start_date",
            "type": "string"
          },
          {
            "description": "End date (YYYY-MM-DD)",
            "in": "query",
            "name": "end_date",
            "type": "string"
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/models.APIResponse"
            }
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ],
        "summary": "[DEPRECATED] Get payments list",
        "tags": [
          "Deprecated-Payments"
        ]
      }
    },
...
    "/api/payments/analytics": {
      "get": {
        "consumes": [
          "application/json"
        ],
        "description": "Get detailed payment analytics for dashboard",
        "parameters": [
          {
            "description": "Start date (YYYY-MM-DD)",
            "in": "query",
            "name": "start_date",
            "required": true,
            "type": "string"
          },
          {
            "description": "End date (YYYY-MM-DD)",
            "in": "query",
            "name": "end_date",
            "required": true,
            "type": "string"
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "additionalProperties": true,
              "type": "object"
            }
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ],
        "summary": "Get payment analytics",
        "tags": [
          "Payments"
        ]
      }
    },
    "/api/payments/debug/recent": {
      "get": {
        "consumes": [
          "application/json"
        ],
        "description": "Get recent payments to verify Sales-Payment integration",
        "parameters": [
          {
            "description": "Hours back to check (default: 24)",
            "in": "query",
            "name": "hours",
            "type": "integer"
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "additionalProperties": true,
              "type": "object"
            }
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ],
        "summary": "Get recent payments for debugging integration",
        "tags": [
          "Payments"
        ]
      }
    },
...
    "/api/payments/export/excel": {
      "get": {
        "consumes": [
          "application/json"
        ],
        "description": "Export payments report in Excel format with filters",
        "parameters": [
          {
            "description": "Start date (YYYY-MM-DD)",
            "in": "query",
            "name": "start_date",
            "type": "string"
          },
          {
            "description": "End date (YYYY-MM-DD)",
            "in": "query",
            "name": "end_date",
            "type": "string"
          },
          {
            "description": "Filter by status",
            "in": "query",
            "name": "status",
            "type": "string"
          },
          {
            "description": "Filter by payment method",
            "in": "query",
            "name": "method",
            "type": "string"
          }
        ],
        "produces": [
          "application/vnd.openxmlformats-officedocument.spreadsheetml.sheet"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "file"
            }
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ],
        "summary": "Export payment report as Excel",
        "tags": [
          "Payments"
        ]
      }
    },
...
    "/api/payments/payable": {
      "post": {
        "consumes": [
          "application/json"
        ],
        "deprecated": true,
        "description": "DEPRECATED: This endpoint may cause double posting. Use SSOT Payment routes instead. Create payment to vendor (payable)",
        "parameters": [
          {
            "description": "Payment data",
            "in": "body",
            "name": "payment",
            "required": true,
            "schema": {
              "$ref": "#/definitions/services.PaymentCreateRequest"
            }
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "201": {
            "description": "Created",
            "schema": {
              "$ref": "#/definitions/models.Payment"
            }
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ],
        "summary": "[DEPRECATED] Create payable payment",
        "tags": [
          "Deprecated-Payments"
        ]
      }
    },
...
    "/api/payments/receivable": {
      "post": {
        "consumes": [
          "application/json"
        ],
        "deprecated": true,
        "description": "DEPRECATED: This endpoint may cause double posting. Use SSOT Payment routes instead. Create payment from customer (receivable)",
        "parameters": [
          {
            "description": "Payment data",
            "in": "body",
            "name": "payment",
            "required": true,
            "schema": {
              "$ref": "#/definitions/services.PaymentCreateRequest"
            }
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "201": {
            "description": "Created",
            "schema": {
              "$ref": "#/definitions/models.Payment"
            }
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ],
        "summary": "[DEPRECATED] Create receivable payment",
        "tags": [
          "Deprecated-Payments"
        ]
      }
    },
    "/api/payments/report/pdf": {
      "get": {
        "consumes": [
          "application/json"
        ],
        "description": "Export payments report in PDF format with filters",
        "parameters": [
          {
            "description": "Start date (YYYY-MM-DD)",
            "in": "query",
            "name": "start_date",
            "type": "string"
          },
          {
            "description": "End date (YYYY-MM-DD)",
            "in": "query",
            "name": "end_date",
            "type": "string"
          },
          {
            "description": "Filter by status",
            "in": "query",
            "name": "status",
            "type": "string"
          },
          {
            "description": "Filter by payment method",
            "in": "query",
            "name": "method",
            "type": "string"
          }
        ],
        "produces": [
          "application/pdf"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "file"
            }
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ],
        "summary": "Export payment report as PDF",
        "tags": [
          "Payments"
        ]
      }
    },
    "/api/payments/summary": {
      "get": {
        "consumes": [
          "application/json"
        ],
        "description": "Get payment summary statistics",
        "parameters": [
          {
            "description": "Start date (YYYY-MM-DD)",
            "in": "query",
            "name": "start_date",
            "required": true,
            "type": "string"
          },
          {
            "description": "End date (YYYY-MM-DD)",
            "in": "query",
            "name": "end_date",
            "required": true,
            "type": "string"
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "additionalProperties": true,
              "type": "object"
            }
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ],
        "summary": "Get payment summary",
        "tags": [
          "Payments"
        ]
      }
    },
    "/api/payments/unpaid-bills/{vendor_id}": {
      "get": {
        "consumes": [
          "application/json"
        ],
        "description": "Get list of unpaid bills for a vendor",
        "parameters": [
          {
            "description": "Vendor ID",
            "in": "path",
            "name": "vendor_id",
            "required": true,
            "type": "integer"
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "items": {
                "$ref": "#/definitions/models.Purchase"
              },
              "type": "array"
            }
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ],
        "summary": "Get unpaid bills",
        "tags": [
          "Payments"
        ]
      }
    },
    "/api/payments/unpaid-invoices/{customer_id}": {
      "get": {
        "consumes": [
          "application/json"
        ],
        "description": "Get list of unpaid invoices for a customer",
        "parameters": [
          {
            "description": "Customer ID",
            "in": "path",
            "name": "customer_id",
            "required": true,
            "type": "integer"
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "items": {
                "$ref": "#/definitions/models.Sale"
              },
              "type": "array"
            }
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ],
        "summary": "Get unpaid invoices",
        "tags": [
          "Payments"
        ]
      }
    },
    "/api/payments/{id}": {
      "delete": {
        "consumes": [
          "application/json"
        ],
        "description": "Delete payment and reverse if needed (Admin Only)",
        "parameters": [
          {
            "description": "Payment ID",
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer"
          },
          {
            "description": "Deletion reason",
            "in": "body",
            "name": "reason",
            "required": true,
            "schema": {
              "additionalProperties": {
                "type": "string"
              },
              "type": "object"
            }
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "additionalProperties": {
                "type": "string"
              },
              "type": "object"
            }
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ],
        "summary": "Delete payment",
        "tags": [
          "Payments"
        ]
      },
      "get": {
        "consumes": [
          "application/json"
        ],
        "deprecated": true,
        "description": "DEPRECATED: This endpoint may cause double posting. Use SSOT Payment routes instead. Get single payment details",
        "parameters": [
          {
            "description": "Payment ID",
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer"
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/models.Payment"
            }
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ],
        "summary": "[DEPRECATED] Get payment by ID",
        "tags": [
          "Deprecated-Payments"
        ]
      }
    },
...
    "/api/payments/{id}/pdf": {
      "get": {
        "consumes": [
          "application/json"
        ],
        "description": "Export single payment details in PDF format",
        "parameters": [
          {
            "description": "Payment ID",
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer"
          }
        ],
        "produces": [
          "application/pdf"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "file"
            }
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ],
        "summary": "Export payment detail as PDF",
        "tags": [
          "Payments"
        ]
      }
    },
...
  },
  "schemes": [
    "http"
  ],
  "securityDefinitions": {
    "BearerAuth": {
      "description": "Type \"Bearer\" followed by a space and JWT token.",
      "in": "header",
      "name": "Authorization",
      "type": "apiKey"
    }
  },
  "swagger": "2.0"
}